+++ network/cache/rcvcache_test.go
@@ -6,7 +6,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/onflow/flow-go/network"
+	"github.com/onflow/flow-go/network/message"
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
@@ -43,20 +43,20 @@ func (r *ReceiveCacheTestSuite) SetupTest() {
 
 // TestSingleElementAdd adds a single element to the cache and verifies its existence.
 func (r *ReceiveCacheTestSuite) TestSingleElementAdd() {
-	eventID, err := network.EventId(channels.Channel("0"), []byte("event-1"))
+	eventID, err := message.EventId(channels.Channel("0"), []byte("event-1"))
 	require.NoError(r.T(), err)
 
 	assert.True(r.Suite.T(), r.c.Add(eventID))
 	assert.False(r.Suite.T(), r.c.Add(eventID))
 
 	// same channel but different event should be treated as unseen
-	eventID2, err := network.EventId(channels.Channel("0"), []byte("event-2"))
+	eventID2, err := message.EventId(channels.Channel("0"), []byte("event-2"))
 	require.NoError(r.T(), err)
 	assert.True(r.Suite.T(), r.c.Add(eventID2))
 	assert.False(r.Suite.T(), r.c.Add(eventID2))
 
 	// same event but different channels should be treated as unseen
-	eventID3, err := network.EventId(channels.Channel("1"), []byte("event-2"))
+	eventID3, err := message.EventId(channels.Channel("1"), []byte("event-2"))
 	require.NoError(r.T(), err)
 	assert.True(r.Suite.T(), r.c.Add(eventID3))
 	assert.False(r.Suite.T(), r.c.Add(eventID3))
@@ -64,7 +64,7 @@ func (r *ReceiveCacheTestSuite) TestSingleElementAdd() {
 
 // TestNoneExistence evaluates the correctness of cache operation against non-existing element
 func (r *ReceiveCacheTestSuite) TestNoneExistence() {
-	eventID, err := network.EventId(channels.Channel("1"), []byte("non-existing event"))
+	eventID, err := message.EventId(channels.Channel("1"), []byte("non-existing event"))
 	require.NoError(r.T(), err)
 
 	// adding new event to cache should return true
@@ -76,7 +76,7 @@ func (r *ReceiveCacheTestSuite) TestMultipleElementAdd() {
 	// creates and populates slice of 10 events
 	eventIDs := make([]hash.Hash, 0)
 	for i := 0; i < r.size; i++ {
-		eventID, err := network.EventId(channels.Channel("1"), []byte(fmt.Sprintf("event-%d", i)))
+		eventID, err := message.EventId(channels.Channel("1"), []byte(fmt.Sprintf("event-%d", i)))
 		require.NoError(r.T(), err)
 
 		eventIDs = append(eventIDs, eventID)
@@ -114,7 +114,7 @@ func (r *ReceiveCacheTestSuite) TestLRU() {
 	eventIDs := make([]hash.Hash, 0)
 	total := r.size + 1
 	for i := 0; i < total; i++ {
-		eventID, err := network.EventId(channels.Channel("1"), []byte(fmt.Sprintf("event-%d", i)))
+		eventID, err := message.EventId(channels.Channel("1"), []byte(fmt.Sprintf("event-%d", i)))
 		require.NoError(r.T(), err)
 
 		eventIDs = append(eventIDs, eventID)
