+++ network/test/middleware_test.go
@@ -748,7 +748,7 @@ func (m *MiddlewareTestSuite) TestEcho() {
 			require.Equal(m.T(), message.ProtocolTypeUnicast, msg.Protocol())                             // protocol
 			require.Equal(m.T(), expectedSendMsg, msg.DecodedPayload().(*libp2pmessage.TestMessage).Text) // payload
 			// event id
-			eventId, err := network.EventId(msg.Channel(), msg.Proto().Payload)
+			eventId, err := message.EventId(msg.Channel(), msg.Proto().Payload)
 			require.NoError(m.T(), err)
 			require.True(m.T(), bytes.Equal(eventId, msg.EventID()))
 
@@ -771,7 +771,7 @@ func (m *MiddlewareTestSuite) TestEcho() {
 			require.Equal(m.T(), message.ProtocolTypeUnicast, msg.Protocol())                              // protocol
 			require.Equal(m.T(), expectedReplyMsg, msg.DecodedPayload().(*libp2pmessage.TestMessage).Text) // payload
 			// event id
-			eventId, err := network.EventId(msg.Channel(), msg.Proto().Payload)
+			eventId, err := message.EventId(msg.Channel(), msg.Proto().Payload)
 			require.NoError(m.T(), err)
 			require.True(m.T(), bytes.Equal(eventId, msg.EventID()))
 		})
@@ -853,7 +853,7 @@ func (m *MiddlewareTestSuite) TestLargeMessageSize_SendDirect() {
 			require.Equal(m.T(), targetNode, msg.TargetIDs()[0])
 			require.Equal(m.T(), message.ProtocolTypeUnicast, msg.Protocol())
 
-			eventId, err := network.EventId(msg.Channel(), msg.Proto().Payload)
+			eventId, err := message.EventId(msg.Channel(), msg.Proto().Payload)
 			require.NoError(m.T(), err)
 			require.True(m.T(), bytes.Equal(eventId, msg.EventID()))
 			close(ch)
